+++ CHANGELOG.md
@@ -15,4 +15,5 @@ the [releases page](https://github.com/Consensys/teku/releases).
 - Added `--ee-jwt-claim-id` command line option to provide `id` to the execution engine JWT claims
 
 ### Bug Fixes
+- Fixed the command line help not displaying `--checkpoint-sync-url` as an option.
 - Fixed bug preventing node to startup when using `--exit-when-no-validator-keys-enabled` even with keys present